TRANSFORM/DATABASE/requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
tqdm==4.66.1
